move util-like things to utils
