+++ pool/stats/tests/test_selenium.py
@@ -4,8 +4,9 @@ from django.urls import reverse
 from selenium.webdriver.support.ui import Select
 
 from .base_cases import BaseSeleniumPoolStatsTestCase, form_length_map, location_names
-from ..models import ScoreSheet, Team, PlayerSeasonSummary
-from ..admin import expire_page
+
+from pool.stats.models import ScoreSheet, Team, PlayerSeasonSummary
+from pool.stats.utils import expire_page
 
 
 class BaseViewRedirectTestCase(BaseSeleniumPoolStatsTestCase):
